+++ src/LnskyDB/QueryFactory.cs
@@ -19,7 +19,7 @@ namespace LnskyDB
             T obj = new T();
             return Create(obj);
         }
-        public static IQuery<T> Create<T>(T obj) where T : BaseDBModel, new()
+        private static IQuery<T> Create<T>(T obj) where T : BaseDBModel, new()
         {
             return new QueryInfo<T>(obj);
         }
